requirements.txt
streamlit>=1.12.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-image>=0.18.0
pillow>=8.0.0